Avoid warning about type conversion.

+++ src/libxsmm_malloc.c
@@ -1001,7 +1001,7 @@ LIBXSMM_API_DEFINITION void* libxsmm_aligned_scratch(size_t size, size_t alignme
           LIBXSMM_LOCK_ACQUIRE(&libxsmm_lock_global);
           if (0 == internal_malloc_scratch_pool[i].buffer) {
             const double scratch_scale = 0 < libxsmm_scratch_scale ? libxsmm_scratch_scale : 1.4;
-            const size_t minsize = scratch_scale * LIBXSMM_MAX(internal_malloc_scratch_pool[i].minsize, req_size);
+            const size_t minsize = (size_t)(scratch_scale * LIBXSMM_MAX(internal_malloc_scratch_pool[i].minsize, req_size));
             assert(0 == internal_malloc_scratch_pool[i].head/*sanity check*/);
             if (EXIT_SUCCESS == libxsmm_xmalloc((void**)&internal_malloc_scratch_pool[i].buffer, minsize, 0/*auto*/,
               LIBXSMM_MALLOC_FLAG_SCRATCH, 0/*extra*/, 0/*extra_size*/))
